add search filter to student list

+++ app/Http/Controllers/GradeController.php
@@ -41,12 +41,14 @@ class GradeController extends Controller
         ]);
 
         return redirect()->route('grades.index')->with('success', 'پایه با موفقیت ویرایش شد.');
-
-      
     }
 
     public function destroy(Grade $grade)
     {
+        // بررسی اینکه آیا دانش‌آموزی به این رشته متصل است
+        if ($grade->students()->exists()) {
+            return back()->with('error', 'این پایه قابل حذف نیست، زیرا دانش‌آموزانی به آن متصل هستند.');
+        }
         $grade->delete();
         return back()->with('success', 'پایه حذف شد.');
     }
